Optimizing performance of ElemElemGraph Updating.
Add early-exit to the get_ordinal_and_permutation function.

i.e., exit the loop over sides early if you've already found a match.

Hotspot shown in vtune for the vivace/p050 performance test.

Only enter ElemElemGraph::delete_elements if there are actually elements
to delete.

This fixes a huge performance issue in the vivace/p050 performance
test.

Add early-exit optimizations for coincident element detection.

Only count all shared sides for a graph-edge if the graph has num-edges
greater than the number of topology sides for that element. If less or
equal, there are cheaper checks that can be made.

Helps fix another hotspot shown in vtune for vivace p050 performance test.

(From STK/SEACAS)

+++ packages/stk/stk_mesh/stk_mesh/baseImpl/elementGraph/ElemElemGraphUpdater.hpp
@@ -57,8 +57,13 @@ public:
 
     virtual void started_modification_end_notification()
     {
-        elemGraph.delete_elements(elementsDeleted);
-        elementsDeleted.clear();
+        size_t numLocalElementsDeleted = elementsDeleted.size();
+        size_t numGlobalElementsDeleted = 0;
+        stk::all_reduce_sum(bulkData.parallel(), &numLocalElementsDeleted, &numGlobalElementsDeleted, 1);
+        if (numGlobalElementsDeleted > 0) {
+            elemGraph.delete_elements(elementsDeleted);
+            elementsDeleted.clear();
+        }
     }
 
     virtual void elements_about_to_move_procs_notification(const stk::mesh::EntityProcVec &elemProcPairsToMove)
